Implement a new descriptor-based decoder
Implementation of a new decoding structure which makes it much easier to
add new instructions and therefore to maintain while removing many lines
of code. Each instruction is now only a descriptor specifying the
instruction type, the opcodes, the operand types and the encoding. The
descriptors are in a separate file to allow usage in other places, e.g.
the generator.

Beside the descriptors, also custom decoding handlers are supported.